backend/requirements.txt
fastapi~=0.115.8
uvicorn~=0.34.0
pydantic~=2.10.6
sqlalchemy~=2.0.31
psycopg2-binary~=2.9.9
requests~=2.32.3
beautifulsoup4~=4.13.3
selenium~=4.29.0
# webdriver-manager~=4.0.2 # Often installed globally or managed differently
Unidecode~=1.3.8
ujson~=5.10.0
orjson~=3.10.15
attrs~=25.1.0
python-dateutil~=2.9.0.post0
starlette~=0.45.3 # Usually a dependency of FastAPI
# gspread~=6.2.0 # For Google Sheets, uncomment if needed
# oauth2client~=4.1.3 # For Google Sheets, uncomment if needed
# google-api-python-client~=2.162.0 # For Google services, uncomment if needed
openai~=1.66.3
python-dotenv~=1.0.1
python-jose[cryptography]~=3.3.0
passlib[bcrypt]~=1.7.4
python-multipart~=0.0.9
alembic~=1.13.1
email-validator~=2.1.0.post1 